+++ tests/integration/conftest.py
@@ -1,19 +1,26 @@
 #!/usr/bin/env python3
 # Copyright 2024 Canonical Ltd.
 # See LICENSE file for licensing details.
+import asyncio
 import logging
 import os
 import subprocess
 from pathlib import Path
-from typing import Iterable
+from typing import Any, AsyncGenerator, Callable, Coroutine, Dict, Generator, Iterable
 
 import boto3
 import boto3.session
 import jubilant
 import pytest
+import pytest_asyncio
 from botocore.client import Config
 from dotenv import load_dotenv
+from lightkube import Client, KubeConfig
+from playwright.async_api import async_playwright
+from playwright.async_api._generated import Browser, BrowserContext, BrowserType, Page
+from playwright.async_api._generated import Playwright as AsyncPlaywright
 
+from .oauth_tools.external_idp import DexIdpService
 from .types import AzureInfo, CharmVersion, IntegrationTestsCharms, S3Info
 
 load_dotenv("microceph.source")
@@ -23,6 +30,7 @@ load_dotenv()
 logger = logging.getLogger(__name__)
 BUCKET_NAME = "history-server"
 PATH_NAME = "spark-events"
+KUBECONFIG = os.environ.get("TESTING_KUBECONFIG", "~/.kube/config")
 
 
 @pytest.fixture(scope="module")
@@ -222,3 +230,112 @@ def history_server_charm() -> Path:
         raise FileNotFoundError("Could not find packed history server charm.")
 
     return path
+
+
+@pytest.fixture(scope="session")
+def client() -> Client:
+    return Client(config=KubeConfig.from_file(KUBECONFIG), field_manager="dex-test")
+
+
+@pytest.fixture(scope="module")
+def event_loop():
+    loop = asyncio.get_event_loop()
+    yield loop
+    loop.close()
+
+
+@pytest.fixture(scope="module")
+def external_idp_service(
+    request: pytest.FixtureRequest, client: Client
+) -> Generator[DexIdpService, None, None]:
+    """Deploy and manage the lifecycle of an Dex service."""
+    logger.info("Deploying dex resources")
+    ext_idp_manager = DexIdpService(client=client)
+    try:
+        yield ext_idp_manager
+    finally:
+        keep_models = bool(request.config.getoption("--keep-models"))
+        if keep_models:
+            return
+        logger.info("Deleting dex resources")
+        ext_idp_manager.remove_idp_service()
+
+
+@pytest.fixture(scope="module")
+def launch_arguments(pytestconfig: Any) -> Dict:
+    return {
+        "headless": not (pytestconfig.getoption("--headed") or os.getenv("HEADFUL", False)),
+        "channel": pytestconfig.getoption("--browser-channel"),
+    }
+
+
+@pytest_asyncio.fixture(scope="module")
+async def playwright() -> AsyncGenerator[AsyncPlaywright, None]:
+    async with async_playwright() as playwright_object:
+        yield playwright_object
+
+
+@pytest.fixture(scope="module")
+def browser_type(playwright: AsyncPlaywright, browser_name: str) -> BrowserType:
+    if browser_name == "firefox":
+        return playwright.firefox
+    if browser_name == "webkit":
+        return playwright.webkit
+    return playwright.chromium
+
+
+@pytest_asyncio.fixture(scope="module")
+async def browser_factory(
+    launch_arguments: Dict, browser_type: BrowserType
+) -> AsyncGenerator[Callable[..., Coroutine[Any, Any, Browser]], None]:
+    browsers = []
+
+    async def launch(**kwargs: Any) -> Browser:
+        browser = await browser_type.launch(**launch_arguments, **kwargs)
+        browsers.append(browser)
+        return browser
+
+    yield launch
+    for browser in browsers:
+        await browser.close()
+
+
+@pytest_asyncio.fixture(scope="module")
+async def browser(
+    browser_factory: Callable[..., Coroutine[Any, Any, Browser]],
+) -> AsyncGenerator[Browser, None]:
+    browser = await browser_factory()
+    yield browser
+    await browser.close()
+
+
+@pytest_asyncio.fixture
+async def context_factory(
+    browser: Browser,
+) -> AsyncGenerator[Callable[..., Coroutine[Any, Any, BrowserContext]], None]:
+    contexts = []
+
+    async def launch(**kwargs: Any) -> BrowserContext:
+        context = await browser.new_context(**kwargs)
+        contexts.append(context)
+        return context
+
+    yield launch
+    for context in contexts:
+        await context.close()
+
+
+@pytest_asyncio.fixture
+async def context(
+    context_factory: Callable[..., Coroutine[Any, Any, BrowserContext]],
+) -> AsyncGenerator[BrowserContext, None]:
+    context = await context_factory(ignore_https_errors=True)
+    yield context
+    await context.close()
+
+
+@pytest_asyncio.fixture
+async def page(context: BrowserContext) -> AsyncGenerator[Page, None]:
+    page = await context.new_page()
+    yield page
+    await page.close()
