+++ tests/integration/test_oauth2proxy.py
@@ -2,7 +2,6 @@
 # Copyright 2025 Canonical Limited
 # See LICENSE file for licensing details.
 
-import asyncio
 import json
 import logging
 import os
@@ -10,19 +9,14 @@ import subprocess
 import urllib.request
 from pathlib import Path
 from time import sleep
-from typing import Any, AsyncGenerator, Callable, Coroutine, Dict, Generator, Optional
+from typing import Optional
 
 import jubilant
-import pytest
-import pytest_asyncio
 import requests
 import yaml
-from lightkube import Client, KubeConfig
-from playwright.async_api import async_playwright
-from playwright.async_api._generated import Browser, BrowserContext, BrowserType, Page
-from playwright.async_api._generated import Playwright as AsyncPlaywright
+from playwright.async_api._generated import BrowserContext, Page
 
-from .oauth_tools.external_idp import DexIdpService, ExternalIdpService
+from .oauth_tools.external_idp import ExternalIdpService
 from .test_helpers import (
     set_s3_credentials,
 )
@@ -34,26 +28,19 @@ METADATA = yaml.safe_load(Path("./metadata.yaml").read_text())
 APP_NAME = METADATA["name"]
 BUCKET_NAME = "history-server"
 
-KUBECONFIG = os.environ.get("TESTING_KUBECONFIG", "~/.kube/config")
+# KUBECONFIG = os.environ.get("TESTING_KUBECONFIG", "~/.kube/config")
 
 
-@pytest.fixture(scope="session")
-def client() -> Client:
-    return Client(config=KubeConfig.from_file(KUBECONFIG), field_manager="dex-test")
+# @pytest.fixture(scope="session")
+# def client() -> Client:
+#     return Client(config=KubeConfig.from_file(KUBECONFIG), field_manager="dex-test")
 
 
-# @pytest_asyncio.fixture
-# async def page(context: BrowserContext) -> AsyncGenerator[Page, None]:
-#     page = await context.new_page()
-#     yield page
-#     await page.close()
-
-
-@pytest.fixture(scope="module")
-def event_loop():
-    loop = asyncio.get_event_loop()
-    yield loop
-    loop.close()
+# @pytest.fixture(scope="module")
+# def event_loop():
+#     loop = asyncio.get_event_loop()
+#     yield loop
+#     loop.close()
 
 
 async def verify_page_loads(page: Page, url: str):
@@ -93,29 +80,6 @@ async def get_cookie_from_browser_by_name(
     return None
 
 
-async def access_application_login_page(
-    page: Page, url: str, redirect_login_url: Optional[str] = None
-):
-    """Navigate the browser to the login page.
-
-    If the url of the application redirects to a login page, pass the application's url as url,
-    and a pattern string for the login page as redirect_login_url.
-    Otherwise pass the url of the application's login page as url, and leave redirect_login_url
-    empty.
-
-    Args:
-        page (page): The page fixture.
-        url (str): The url to go to.
-        redirect_login_url (str): The redirect to which the browser will get redirected to.
-    """
-    await page.goto(url)
-    logger.info(f"Navigated to {url}")
-    logger.info(f"Page URL after navigation: {page.url}")
-    logger.info(f"Page content after navigation: {await page.content()}")
-    # if redirect_login_url:
-    #     await expect(page).to_have_url(re.compile(rf"{redirect_login_url}*"))
-
-
 async def complete_auth_code_login(
     page: Page,
     external_idp_service: Optional[ExternalIdpService],
@@ -133,101 +97,101 @@ async def complete_auth_code_login(
     # logger.info(f"Page content after login flow: {await page.content()}")
 
 
-@pytest.fixture(scope="module")
-def external_idp_service(
-    request: pytest.FixtureRequest, client: Client
-) -> Generator[DexIdpService, None, None]:
-    """Deploy and manage the lifecycle of an Dex service."""
-    logger.info("Deploying dex resources")
-    ext_idp_manager = DexIdpService(client=client)
-    try:
-        yield ext_idp_manager
-    finally:
-        keep_models = bool(request.config.getoption("--keep-models"))
-        if keep_models:
-            return
-        logger.info("Deleting dex resources")
-        ext_idp_manager.remove_idp_service()
+# @pytest.fixture(scope="module")
+# def external_idp_service(
+#     request: pytest.FixtureRequest, client: Client
+# ) -> Generator[DexIdpService, None, None]:
+#     """Deploy and manage the lifecycle of an Dex service."""
+#     logger.info("Deploying dex resources")
+#     ext_idp_manager = DexIdpService(client=client)
+#     try:
+#         yield ext_idp_manager
+#     finally:
+#         keep_models = bool(request.config.getoption("--keep-models"))
+#         if keep_models:
+#             return
+#         logger.info("Deleting dex resources")
+#         ext_idp_manager.remove_idp_service()
 
 
-@pytest.fixture(scope="module")
-def launch_arguments(pytestconfig: Any) -> Dict:
-    return {
-        "headless": not (pytestconfig.getoption("--headed") or os.getenv("HEADFUL", False)),
-        "channel": pytestconfig.getoption("--browser-channel"),
-    }
+# @pytest.fixture(scope="module")
+# def launch_arguments(pytestconfig: Any) -> Dict:
+#     return {
+#         "headless": not (pytestconfig.getoption("--headed") or os.getenv("HEADFUL", False)),
+#         "channel": pytestconfig.getoption("--browser-channel"),
+#     }
 
 
-@pytest_asyncio.fixture(scope="module")
-async def playwright() -> AsyncGenerator[AsyncPlaywright, None]:
-    async with async_playwright() as playwright_object:
-        yield playwright_object
+# @pytest_asyncio.fixture(scope="module")
+# async def playwright() -> AsyncGenerator[AsyncPlaywright, None]:
+#     async with async_playwright() as playwright_object:
+#         yield playwright_object
 
 
-@pytest.fixture(scope="module")
-def browser_type(playwright: AsyncPlaywright, browser_name: str) -> BrowserType:
-    if browser_name == "firefox":
-        return playwright.firefox
-    if browser_name == "webkit":
-        return playwright.webkit
-    return playwright.chromium
+# @pytest.fixture(scope="module")
+# def browser_type(playwright: AsyncPlaywright, browser_name: str) -> BrowserType:
+#     if browser_name == "firefox":
+#         return playwright.firefox
+#     if browser_name == "webkit":
+#         return playwright.webkit
+#     return playwright.chromium
 
 
-@pytest_asyncio.fixture(scope="module")
-async def browser_factory(
-    launch_arguments: Dict, browser_type: BrowserType
-) -> AsyncGenerator[Callable[..., Coroutine[Any, Any, Browser]], None]:
-    browsers = []
+# @pytest_asyncio.fixture(scope="module")
+# async def browser_factory(
+#     launch_arguments: Dict, browser_type: BrowserType
+# ) -> AsyncGenerator[Callable[..., Coroutine[Any, Any, Browser]], None]:
+#     browsers = []
 
-    async def launch(**kwargs: Any) -> Browser:
-        browser = await browser_type.launch(**launch_arguments, **kwargs)
-        browsers.append(browser)
-        return browser
+#     async def launch(**kwargs: Any) -> Browser:
+#         browser = await browser_type.launch(**launch_arguments, **kwargs)
+#         browsers.append(browser)
+#         return browser
 
-    yield launch
-    for browser in browsers:
-        await browser.close()
-
-
-@pytest_asyncio.fixture(scope="module")
-async def browser(
-    browser_factory: Callable[..., Coroutine[Any, Any, Browser]],
-) -> AsyncGenerator[Browser, None]:
-    browser = await browser_factory()
-    yield browser
-    await browser.close()
+#     yield launch
+#     for browser in browsers:
+#         await browser.close()
 
 
-@pytest_asyncio.fixture
-async def context_factory(
-    browser: Browser,
-) -> AsyncGenerator[Callable[..., Coroutine[Any, Any, BrowserContext]], None]:
-    contexts = []
-
-    async def launch(**kwargs: Any) -> BrowserContext:
-        context = await browser.new_context(**kwargs)
-        contexts.append(context)
-        return context
-
-    yield launch
-    for context in contexts:
-        await context.close()
+# @pytest_asyncio.fixture(scope="module")
+# async def browser(
+#     browser_factory: Callable[..., Coroutine[Any, Any, Browser]],
+# ) -> AsyncGenerator[Browser, None]:
+#     browser = await browser_factory()
+#     yield browser
+#     await browser.close()
 
 
-@pytest_asyncio.fixture
-async def context(
-    context_factory: Callable[..., Coroutine[Any, Any, BrowserContext]],
-) -> AsyncGenerator[BrowserContext, None]:
-    context = await context_factory(ignore_https_errors=True)
-    yield context
-    await context.close()
+# @pytest_asyncio.fixture
+# async def context_factory(
+#     browser: Browser,
+# ) -> AsyncGenerator[Callable[..., Coroutine[Any, Any, BrowserContext]], None]:
+#     contexts = []
+
+#     async def launch(**kwargs: Any) -> BrowserContext:
+#         context = await browser.new_context(**kwargs)
+#         contexts.append(context)
+#         return context
+
+#     yield launch
+#     for context in contexts:
+#         await context.close()
 
 
-@pytest_asyncio.fixture
-async def page(context: BrowserContext) -> AsyncGenerator[Page, None]:
-    page = await context.new_page()
-    yield page
-    await page.close()
+# @pytest_asyncio.fixture
+# async def context(
+#     context_factory: Callable[..., Coroutine[Any, Any, BrowserContext]],
+# ) -> AsyncGenerator[BrowserContext, None]:
+#     context = await context_factory(ignore_https_errors=True)
+#     yield context
+#     await context.close()
+
+
+# @pytest_asyncio.fixture
+# async def page(context: BrowserContext) -> AsyncGenerator[Page, None]:
+#     page = await context.new_page()
+#     yield page
+#     await page.close()
 
 
 def test_build_and_deploy(
